backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firebase-admin>=6.8.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0"
]